# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NHCE Campus Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── POSTS ─────────────────────────────────────────────────────────────────
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List feed posts
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      operationId: createPost
      tags: [posts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    patch:
      operationId: updatePost
      tags: [posts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      operationId: deletePost
      tags: [posts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /posts/{id}/like:
    post:
      operationId: togglePostLike
      tags: [posts]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /posts/{id}/pin:
    patch:
      operationId: pinPost
      tags: [posts]
      summary: Pin/unpin post (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinToggle"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  # ── COMMENTS ──────────────────────────────────────────────────────────────
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── COMPLAINTS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, name, email, role, createdAt]
      properties:
        id:          { type: string }
        clerkId:     { type: string }
        name:        { type: string }
        email:       { type: string }
        role:        { type: string }
        usn:         { type: ["string", "null"] }
        branch:      { type: ["string", "null"] }
        year:        { type: ["integer", "null"] }
        course:      { type: ["string", "null"] }
        avatarUrl:   { type: ["string", "null"] }
        isAdmin:     { type: boolean }
        createdAt:   { type: string }
...
    Post:
      type: object
      required: [id, authorId, content, category, likeCount, commentCount, isPinned, isAnonymous, createdAt]
      properties:
        id:           { type: string }
        authorId:     { type: string }
        authorName:   { type: ["string", "null"] }
        authorAvatar: { type: ["string", "null"] }
        content:      { type: string }
        imageUrl:     { type: ["string", "null"] }
        category:     { type: string }
        likeCount:    { type: integer }
        commentCount: { type: integer }
        isPinned:     { type: boolean }
        isAnonymous:  { type: boolean }
        isLiked:      { type: boolean }
        createdAt:    { type: string }
    PostInput:
      type: object
      required: [content, category]
      properties:
        content:     { type: string, minLength: 1 }
        imageUrl:    { type: string }
        category:    { type: string }
        isAnonymous: { type: boolean }
    PostUpdate:
      type: object
      properties:
        content:  { type: string }
        imageUrl: { type: string }
    PostListResponse:
      type: object
      required: [posts, total]
      properties:
        posts: { type: array, items: { $ref: "#/components/schemas/Post" } }
        total: { type: integer }
    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked:     { type: boolean }
        likeCount: { type: integer }
    PinToggle:
      type: object
      required: [isPinned]
      properties:
        isPinned: { type: boolean }

    # ── COMMENT ─────────────────────────────────────────────────────────────
    Comment:
      type: object
      required: [id, postId, authorId, content, createdAt]
      properties:
        id:           { type: string }
        postId:       { type: string }
        authorId:     { type: string }
        authorName:   { type: ["string", "null"] }
        authorAvatar: { type: ["string", "null"] }
        content:      { type: string }
        createdAt:    { type: string }
    CommentInput:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1 }

    # ── COMPLAINT ───────────────────────────────────────────────────────────
...