# lib/api-zod/src/generated/types/postListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NHCE Campus Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from './post';

export interface PostListResponse {
  posts: Post[];
  total: number;
}